requirements.txt
absl-py==2.1.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
coverage==7.10.2
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.18
enum34==1.1.10
executing==2.0.1
fastapi==0.116.1
filelock==3.16.0
flatbuffers==24.3.25
fonttools==4.53.1
fpdf==1.7.2
fsspec==2024.9.0
gdown==5.2.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.9
imageio==2.35.1
iniconfig==2.0.0
ipython==8.25.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.4.0
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mediapipe==0.10.14
ml-dtypes==0.4.0
mpmath==1.3.0
neo4j==5.25.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
onnxruntime==1.19.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
parso==0.8.4
pdf2image==1.17.0
piexif==1.1.3
pikepdf==9.7.0
pillow==10.4.0
platformdirs==4.3.3
pluggy==1.5.0
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==4.25.4
psycopg2-binary==2.9.10
pure-eval==0.2.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.18.0
PyMatting==1.1.12
pyparsing==3.1.2
pypdf==5.4.0
pyperclip==1.9.0
pyreadline3==3.5.2
PySocks==1.7.1
pytest==8.3.3
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soupsieve==2.6
SQLAlchemy==2.0.42
sqlparse==0.5.0
stack-data==0.6.3
stag==2.1.0
starlette==0.47.2
sympy==1.13.2
threadpoolctl==3.5.0
tifffile==2024.8.30
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.3
transitions==0.9.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.35.0
wcwidth==0.2.13
websocket-client==1.8.0
wheel==0.43.0
wrapt==1.17.2
